[DwarfDebug] Move emission of types from beginModule() to endModule() (7/7)
RFC https://discourse.llvm.org/t/rfc-dwarfdebug-fix-and-improve-handling-imported-entities-types-and-static-local-in-subprogram-and-lexical-block-scopes/68544

This patch proposes to move the emission of types and type-like entities from
`DwarfDebug::beginModule()` to `DwarfDebug::endModule()`. Effectively, this changes
nothing but the order of debug entities in the resulting DWARF.

This is needed to simplify DWARF emission in a context of proper support of
function-local types which comes in the next patch (D144006), making all
the types handled in a single place, together with other global and
local entities.

No functional changes intended.

Differential Revision: https://reviews.llvm.org/D144005

Depends on D144004